refactor(git): isolate selected commit transaction
